+++ src/ChangeLog.md
@@ -1,3 +1,72 @@
+
+## NOT RELEASED: 5.8.0
+
+    Generated metadata - do not edit. This metadata will not appear on a release build.
+    
+    Head commit:           1.006.0
+    Branch:                master
+    Contributing commits:  8
+    
+    Prior releases contributing to this changelog:
+      * 5.7.0
+      * 5.7.1
+      * 5.6.99
+    
+    Generated metadata is used as part of incrementally updating this changelog.
+    Edits made below will be preserved and changes from new commits will be added
+    to this file. Delete this file to force it to be regenerated.
+
+## Added
+
+* added feature.
+
+## Changed
+
+None.
+
+## Fixed
+
+* fixed bug.
+
+
+
+
+# Changelog
+
+All notable changes to this project will be documented in this file.
+                 
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+
+## NOT RELEASED: 5.8.0 - _Saturday, 28 June 2025_
+
+    Generated metadata. Do not edit this section.
+    This will not appear on a release build.
+    
+    Up to head commit 1.006.0 on branch master.
+    8 contributing commits.
+    Prior releases contributing to this changelog:
+      * 5.7.0
+      * 5.7.1
+      * 5.6.99
+
+
+## Added
+
+* added feature.
+
+## Changed
+
+None.
+
+## Fixed
+
+* fixed bug.
+
+
+
 # Git2SemVer.MSBuild Changelog
 
 All notable changes to this project will be documented in this file.